fix(lib-cloudinfo): type CloudProduct.gpusPerVm as Float to keep fractional-GPU counts
CloudInfo reports fractional GPU counts for fractional-GPU instance families
(e.g. AWS g6f.2xlarge reports gpusPerVm=0.25). The Integer field truncated
these to 0 on JSON decode, making genuine GPU instances indistinguishable from
CPU-only ones for consumers that check gpusPerVm > 0.

Type gpusPerVm as Float (consistent with memPerVm/onDemandPrice) so fractional
counts survive; > 0 then correctly flags any accelerator.

Fixes #97

+++ lib-cloudinfo/src/main/java/io/seqera/cloudinfo/api/CloudProduct.java
@@ -37,7 +37,16 @@ public class CloudProduct {
     private String category;
     private Integer cpusPerVm;
     private Float memPerVm;
-    private Integer gpusPerVm;
+    /**
+     * Number of GPUs (accelerators) per instance, as reported by CloudInfo.
+     * Modelled as a {@link Float} because fractional-GPU families report a
+     * fraction of a physical accelerator (e.g. AWS {@code g6f.2xlarge} reports
+     * {@code 0.25}); an integer field would truncate such values to {@code 0}
+     * and misclassify the instance as CPU-only. A value {@code > 0} (including
+     * a fraction) means the instance carries an accelerator. Null on legacy
+     * responses that do not populate it.
+     */
+    private Float gpusPerVm;
     private Boolean currentGen;
     private String ntwPerf;
     private String ntwPerfCategory;
@@ -101,11 +110,11 @@ public class CloudProduct {
         this.memPerVm = memPerVm;
     }
 
-    public Integer getGpusPerVm() {
+    public Float getGpusPerVm() {
         return gpusPerVm;
     }
 
-    public void setGpusPerVm(Integer gpusPerVm) {
+    public void setGpusPerVm(Float gpusPerVm) {
         this.gpusPerVm = gpusPerVm;
     }
